refactor: impl 디렉터리 삭제로 상위 디렉터리로 서비스 옮김
