+++ src/main/java/com/psj/itembrowser/cart/controller/CartApiController.java
@@ -16,7 +16,7 @@ import com.psj.itembrowser.cart.domain.dto.request.CartProductDeleteRequestDTO;
 import com.psj.itembrowser.cart.domain.dto.request.CartProductRequestDTO;
 import com.psj.itembrowser.cart.domain.dto.request.CartProductUpdateRequestDTO;
 import com.psj.itembrowser.cart.domain.dto.response.CartResponseDTO;
-import com.psj.itembrowser.cart.service.impl.CartService;
+import com.psj.itembrowser.cart.service.CartService;
 import com.psj.itembrowser.security.common.message.MessageDTO;
 
 import lombok.RequiredArgsConstructor;
