+++ src/main/java/com/psj/itembrowser/order/service/AddressValidationResult.java
@@ -1,4 +1,4 @@
-package com.psj.itembrowser.order.service.impl;
+package com.psj.itembrowser.order.service;
 
 import com.psj.itembrowser.security.common.openfeign.dto.AddressResponseDto;
