+++ src/main/java/com/psj/itembrowser/order/service/OrderCalculationService.java
@@ -1,4 +1,4 @@
-package com.psj.itembrowser.order.service.impl;
+package com.psj.itembrowser.order.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -7,7 +7,7 @@ import com.psj.itembrowser.member.domain.vo.Member;
 import com.psj.itembrowser.order.domain.dto.request.OrderCreateRequestDTO;
 import com.psj.itembrowser.order.domain.vo.OrdersProductRelationResponseDTO;
 import com.psj.itembrowser.product.domain.vo.Product;
-import com.psj.itembrowser.product.service.impl.ProductService;
+import com.psj.itembrowser.product.service.ProductService;
 import com.psj.itembrowser.security.common.exception.BadRequestException;
 import com.psj.itembrowser.security.common.exception.ErrorCode;
 
@@ -23,37 +23,37 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class OrderCalculationService {
-	
+
 	private final ProductService productService;
 	private final PercentageDiscountService percentageDiscountService;
 	private final ShippingPolicyService shippingPolicyService;
-	
+
 	public OrderCalculationResult calculateOrderDetails(@NonNull OrderCreateRequestDTO orderCreateRequestDTO, @NonNull Member member) {
 		validateOrderProduct(orderCreateRequestDTO);
-		
+
 		double totalPrice = 0;
 		double totalDiscount = 0;
 		double shippingFee = 0;
-		
+
 		for (OrdersProductRelationResponseDTO ordersProductRelationResponseDTO : orderCreateRequestDTO.getProducts()) {
 			Product product = Product.from(productService.getProduct(ordersProductRelationResponseDTO.getProductId()));
-			
+
 			double productPrice = product.calculateTotalPrice();
-			
+
 			totalPrice += productPrice;
-			
+
 			double discount = percentageDiscountService.calculateDiscount(product, member);
-			
+
 			totalDiscount += discount;
 		}
-		
+
 		shippingFee = shippingPolicyService.getCurrentShippingPolicy().calculateShippingFee(totalPrice).getFee();
-		
+
 		double orderTotal = totalPrice - totalDiscount + shippingFee;
-		
+
 		return OrderCalculationResult.of(totalPrice, totalDiscount, shippingFee, orderTotal);
 	}
-	
+
 	private static void validateOrderProduct(OrderCreateRequestDTO orderCreateRequestDTO) {
 		if (orderCreateRequestDTO.getProducts().isEmpty()) {
 			throw new BadRequestException(ErrorCode.ORDER_PRODUCTS_EMPTY);
