+++ src/main/java/com/psj/itembrowser/product/service/FileUtil.java
@@ -1,6 +1,5 @@
-package com.psj.itembrowser.product.service.impl;
+package com.psj.itembrowser.product.service;
 
-import java.util.Arrays;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
