+++ src/main/java/com/psj/itembrowser/product/service/ProductService.java
@@ -1,4 +1,4 @@
-package com.psj.itembrowser.product.service.impl;
+package com.psj.itembrowser.product.service;
 
 import java.util.List;
 
@@ -12,7 +12,6 @@ import com.psj.itembrowser.product.domain.dto.request.ProductUpdateDTO;
 import com.psj.itembrowser.product.domain.dto.response.ProductResponseDTO;
 import com.psj.itembrowser.product.domain.vo.Product;
 import com.psj.itembrowser.product.persistence.ProductPersistence;
-import com.psj.itembrowser.product.service.FileService;
 
 import lombok.RequiredArgsConstructor;
 
@@ -25,21 +24,21 @@ import lombok.RequiredArgsConstructor;
 @Service
 @RequiredArgsConstructor
 public class ProductService {
-	
+
 	private final ProductPersistence productPersistence;
 	private final FileService fileService;
-	
+
 	@Transactional
 	public boolean modifyProductQuantity(
 		ProductQuantityUpdateRequestDTO productQuantityUpdateRequestDTO) {
 		return productPersistence.updateProductQuantity(productQuantityUpdateRequestDTO);
 	}
-	
+
 	@Transactional(readOnly = true)
 	public ProductResponseDTO getProduct(Long productId) {
 		return ProductResponseDTO.from(productPersistence.findProductById(productId));
 	}
-	
+
 	/**
 	 * @deprecated
 	 */
@@ -48,46 +47,46 @@ public class ProductService {
 	public List<Product> getProducts(Long orderId) {
 		return productPersistence.findProductsByOrderId(orderId);
 	}
-	
+
 	@Transactional
 	public void createProduct(ProductRequestDTO productRequestDTO) {
 		Product product = Product.from(productRequestDTO);
 		List<MultipartFile> files = productRequestDTO.getMultipartFiles();
-		
+
 		product.validateSellDates();
-		
+
 		productPersistence.createProduct(product);
-		
+
 		Long productId = product.getId();
-		
+
 		fileService.createProductImages(files, productId);
 	}
-	
+
 	@Transactional
 	public void updateProduct(ProductUpdateDTO productUpdateDTO, Long productId) {
 		// Ensure data consistency using pessimistic locking
 		Product product = Product.from(productUpdateDTO);
-		
+
 		productPersistence.findProductStatusForUpdate(productId);
-		
+
 		product.validateSellDates();
-		
+
 		productPersistence.updateProduct(product);
-		
+
 		fileService.updateProductImages(productUpdateDTO, productId);
 	}
-	
+
 	public void decreaseStock(Product quantity) {
 		//TODO 재고 관련 로직 추가
 	}
-	
+
 	@Transactional
 	public void deleteProduct(Long productId) {
 		// Ensure data consistency using pessimistic locking
 		productPersistence.findProductStatusForUpdate(productId);
-		
+
 		productPersistence.softDeleteProduct(productId);
-		
+
 		fileService.deleteProductImages(productId);
 	}
 }
