+++ src/test/java/com/psj/itembrowser/cart/controller/CartApiControllerTest.java
@@ -40,7 +40,7 @@ import com.psj.itembrowser.cart.domain.dto.request.CartProductRequestDTO;
 import com.psj.itembrowser.cart.domain.dto.request.CartProductUpdateRequestDTO;
 import com.psj.itembrowser.cart.domain.dto.response.CartProductRelationResponseDTO;
 import com.psj.itembrowser.cart.domain.dto.response.CartResponseDTO;
-import com.psj.itembrowser.cart.service.impl.CartService;
+import com.psj.itembrowser.cart.service.CartService;
 import com.psj.itembrowser.security.common.exception.DatabaseOperationException;
 import com.psj.itembrowser.security.common.exception.ErrorCode;
 import com.psj.itembrowser.security.common.exception.NotFoundException;
@@ -56,20 +56,20 @@ import com.psj.itembrowser.security.common.exception.NotFoundException;
 @AutoConfigureMockMvc
 @AutoConfigureRestDocs
 class CartApiControllerTest {
-	
+
 	private final String BASE_URL = "/v1/api/cart";
-	
+
 	private final String USER_EMAIL = "test@test.com";
-	
+
 	@Autowired
 	MockMvc mockMvc;
-	
+
 	@Autowired
 	ObjectMapper objectMapper;
-	
+
 	@MockBean
 	private CartService cartService;
-	
+
 	@BeforeEach
 	public void setUp(WebApplicationContext webApplicationContext,
 		RestDocumentationContextProvider restDocumentation) {
@@ -78,10 +78,10 @@ class CartApiControllerTest {
 			.apply(MockMvcRestDocumentation.documentationConfiguration(restDocumentation))
 			.build();
 	}
-	
+
 	@Nested
 	class SelectTest {
-		
+
 		@Test
 		@DisplayName("장바구니 목록 조회 API 테스트 -> 정상 조회 테스트")
 		void given_GetCart_Expect_Success() throws Exception {
@@ -89,9 +89,9 @@ class CartApiControllerTest {
 			CartResponseDTO cartResponseDTO = new CartResponseDTO();
 			cartResponseDTO.setUserId(USER_EMAIL);
 			cartResponseDTO.setProducts(List.of(CartProductRelationResponseDTO.builder().build()));
-			
+
 			given(cartService.getCart(anyString())).willReturn(cartResponseDTO);
-			
+
 			// when + then
 			mockMvc
 				.perform(
@@ -133,14 +133,14 @@ class CartApiControllerTest {
 					)
 				));
 		}
-		
+
 		@Test
 		@DisplayName("장바구니 빈값 조회시 API 오류 발생")
 		void given_CartIsNull_Expect_Exception() throws Exception {
 			// given
 			given(cartService.getCart(anyString())).willThrow(
 				new NotFoundException(ErrorCode.CART_NOT_FOUND));
-			
+
 			// given + then
 			mockMvc
 				.perform(RestDocumentationRequestBuilders
@@ -172,12 +172,12 @@ class CartApiControllerTest {
 				));
 		}
 	}
-	
+
 	@Nested
 	class InsertTest {
-		
+
 		CartProductRequestDTO mock;
-		
+
 		@BeforeEach
 		void setUp() {
 			// given
@@ -189,7 +189,7 @@ class CartApiControllerTest {
 				.userId(USER_EMAIL)
 				.build();
 		}
-		
+
 		@Test
 		@DisplayName("장바구니에 상품을 올바르게 삽입시, 정상 응답 테스트")
 		void given_CorrectInsertProductInCart_Expect_Success() throws Exception {
@@ -206,7 +206,7 @@ class CartApiControllerTest {
 					"add-cart",
 					preprocessRequest(prettyPrint()),
 					preprocessResponse(prettyPrint()),
-					
+
 					ResourceDocumentation.resource(
 						ResourceSnippetParameters.builder()
 							.tag("cart")
@@ -224,7 +224,7 @@ class CartApiControllerTest {
 					)
 				));
 		}
-		
+
 		@Test
 		@DisplayName("장바구니에 상품을 잘못 삽입시, 오류 응답 테스트")
 		void given_IncorrectInsertProductInCart_Expect_Exception() throws Exception {
@@ -232,7 +232,7 @@ class CartApiControllerTest {
 			doThrow(new DatabaseOperationException(CART_PRODUCT_INSERT_FAIL))
 				.when(cartService)
 				.addCartProduct(mock);
-			
+
 			mockMvc
 				.perform(RestDocumentationRequestBuilders
 					.post(BASE_URL + "/add")
@@ -244,7 +244,7 @@ class CartApiControllerTest {
 					"add-cart-fail",
 					preprocessRequest(prettyPrint()),
 					preprocessResponse(prettyPrint()),
-					
+
 					ResourceDocumentation.resource(
 						ResourceSnippetParameters.builder()
 							.tag("cart")
@@ -267,12 +267,12 @@ class CartApiControllerTest {
 				));
 		}
 	}
-	
+
 	@Nested
 	class UpdateTest {
-		
+
 		CartProductUpdateRequestDTO mock;
-		
+
 		@BeforeEach
 		void setUp() {
 			// given
@@ -283,7 +283,7 @@ class CartApiControllerTest {
 				.quantity(10)
 				.build();
 		}
-		
+
 		@Test
 		@DisplayName("장바구니 상품 수정 API -> 정상 응답 테스트")
 		void given_UpdateProductInCart_Expect_Success() throws Exception {
@@ -300,7 +300,7 @@ class CartApiControllerTest {
 					"update-cart",
 					preprocessRequest(prettyPrint()),
 					preprocessResponse(prettyPrint()),
-					
+
 					ResourceDocumentation.resource(
 						ResourceSnippetParameters.builder()
 							.tag("cart")
@@ -317,7 +317,7 @@ class CartApiControllerTest {
 					)
 				));
 		}
-		
+
 		@Test
 		@DisplayName("장바구니 상품 수정 API -> 장바구니 상품이 존재하지 않을 경우 오류 발생")
 		void given_UpdateProductInCart_Expect_Exception() throws Exception {
@@ -328,12 +328,12 @@ class CartApiControllerTest {
 				.productId(1L)
 				.quantity(10)
 				.build();
-			
+
 			//when -  then
 			doThrow(new NotFoundException(CART_PRODUCT_UPDATE_FAIL))
 				.when(cartService)
 				.modifyCartProduct(any(CartProductUpdateRequestDTO.class));
-			
+
 			mockMvc
 				.perform(RestDocumentationRequestBuilders
 					.put(BASE_URL + "/update")
@@ -346,7 +346,7 @@ class CartApiControllerTest {
 					"update-cart-fail",
 					preprocessRequest(prettyPrint()),
 					preprocessResponse(prettyPrint()),
-					
+
 					ResourceDocumentation.resource(
 						ResourceSnippetParameters.builder()
 							.tag("cart")
@@ -368,12 +368,12 @@ class CartApiControllerTest {
 				));
 		}
 	}
-	
+
 	@Nested
 	class DeleteTest {
-		
+
 		CartProductDeleteRequestDTO mock;
-		
+
 		@BeforeEach
 		void setUp() {
 			// given
@@ -383,7 +383,7 @@ class CartApiControllerTest {
 				.productId(1L)
 				.build();
 		}
-		
+
 		@Test
 		@DisplayName("장바구니에서 상품 삭제 API -> 정상 응답 테스트")
 		void given_DeleteProductInCart_Expect_Success() throws Exception {
@@ -400,7 +400,7 @@ class CartApiControllerTest {
 					"delete-cart",
 					preprocessRequest(prettyPrint()),
 					preprocessResponse(prettyPrint()),
-					
+
 					ResourceDocumentation.resource(
 						ResourceSnippetParameters.builder()
 							.tag("cart")
@@ -416,7 +416,7 @@ class CartApiControllerTest {
 					)
 				));
 		}
-		
+
 		@Test
 		@DisplayName("장바구니 상품 삭제 API -> 장바구니 상품이 존재하지 않을 경우 오류 발생")
 		void given_DeleteProductInCart_Expect_Exception() throws Exception {
@@ -424,7 +424,7 @@ class CartApiControllerTest {
 			doThrow(new NotFoundException(CART_PRODUCT_DELETE_FAIL))
 				.when(cartService)
 				.removeCart(mock);
-			
+
 			mockMvc
 				.perform(RestDocumentationRequestBuilders
 					.delete(BASE_URL + "/delete")
@@ -437,7 +437,7 @@ class CartApiControllerTest {
 					"delete-cart-fail",
 					preprocessRequest(prettyPrint()),
 					preprocessResponse(prettyPrint()),
-					
+
 					ResourceDocumentation.resource(
 						ResourceSnippetParameters.builder()
 							.tag("cart")
