+++ src/test/java/com/psj/itembrowser/cart/service/impl/CartServiceTest.java
@@ -25,6 +25,7 @@ import com.psj.itembrowser.cart.domain.vo.Cart;
 import com.psj.itembrowser.cart.domain.vo.CartProductRelation;
 import com.psj.itembrowser.cart.mapper.CartMapper;
 import com.psj.itembrowser.cart.persistance.CartPersistence;
+import com.psj.itembrowser.cart.service.CartService;
 import com.psj.itembrowser.product.domain.vo.Product;
 import com.psj.itembrowser.security.common.exception.NotFoundException;
