+++ src/test/java/com/psj/itembrowser/order/controller/OrderSelectApiControllerTest.java
@@ -57,7 +57,7 @@ import com.psj.itembrowser.order.domain.dto.response.OrderResponseDTO;
 import com.psj.itembrowser.order.domain.vo.Order;
 import com.psj.itembrowser.order.domain.vo.OrderStatus;
 import com.psj.itembrowser.order.domain.vo.OrdersProductRelation;
-import com.psj.itembrowser.order.service.impl.OrderService;
+import com.psj.itembrowser.order.service.OrderService;
 import com.psj.itembrowser.product.domain.vo.DeliveryFeeType;
 import com.psj.itembrowser.product.domain.vo.Product;
 import com.psj.itembrowser.product.domain.vo.ProductStatus;
@@ -70,22 +70,22 @@ import com.psj.itembrowser.shippingInfos.domain.vo.ShippingInfo;
 @WebMvcTest(OrderApiController.class)
 @AutoConfigureRestDocs
 class OrderSelectApiControllerTest {
-	
+
 	@Autowired
 	private MockMvc mockMvc;
-	
+
 	@MockBean
 	private OrderService orderService;
-	
+
 	@MockBean
 	private UserDetailsServiceImpl userDetailsService;
-	
+
 	@Mock
 	private Jwt jwt;
-	
+
 	private Order expectedOrderWithADMINUser;
 	private Order expectedOrderWithCUSTOMERUser;
-	
+
 	@BeforeEach
 	public void setUp(WebApplicationContext webApplicationContext,
 		RestDocumentationContextProvider restDocumentation) {
@@ -93,7 +93,7 @@ class OrderSelectApiControllerTest {
 			.webAppContextSetup(webApplicationContext)
 			.apply(MockMvcRestDocumentation.documentationConfiguration(restDocumentation))
 			.build();
-		
+
 		Member expectedAdminMember = Member.builder()
 			.memberNo(1L)
 			.credentials(Credentials.builder()
@@ -112,7 +112,7 @@ class OrderSelectApiControllerTest {
 			.birthday(LocalDate.of(1995, 11, 3))
 			.lastLoginDate(LocalDateTime.now())
 			.build();
-		
+
 		Member expectedCustomerMember = Member.builder()
 			.memberNo(1L)
 			.credentials(Credentials.builder()
@@ -131,7 +131,7 @@ class OrderSelectApiControllerTest {
 			.birthday(LocalDate.of(1995, 11, 3))
 			.lastLoginDate(LocalDateTime.now())
 			.build();
-		
+
 		ShippingInfo expectedShppingInfo = new ShippingInfo(1L,
 			1L,
 			"홍길동",
@@ -143,7 +143,7 @@ class OrderSelectApiControllerTest {
 			LocalDateTime.now(),
 			null,
 			null);
-		
+
 		OrdersProductRelation expectedOrderRelation1 = OrdersProductRelation.of(1L, 1L, 1,
 			LocalDateTime.now(),
 			null,
@@ -169,7 +169,7 @@ class OrderSelectApiControllerTest {
 				Collections.emptyList(),
 				Collections.emptyList()
 			));
-		
+
 		OrdersProductRelation expectedOrderRelation2 = OrdersProductRelation.of(2L, 1L, 1,
 			LocalDateTime.now(),
 			null,
@@ -194,7 +194,7 @@ class OrderSelectApiControllerTest {
 				Collections.emptyList(),
 				Collections.emptyList()
 			));
-		
+
 		this.expectedOrderWithADMINUser = Order.of(
 			1L,
 			1L,
@@ -210,7 +210,7 @@ class OrderSelectApiControllerTest {
 			),
 			expectedAdminMember,
 			expectedShppingInfo);
-		
+
 		this.expectedOrderWithCUSTOMERUser = Order.of(
 			1L,
 			1L,
@@ -227,7 +227,7 @@ class OrderSelectApiControllerTest {
 			expectedCustomerMember,
 			expectedShppingInfo);
 	}
-	
+
 	@Test
 	@MockMember
 	@DisplayName("권한 - CUSTOMER 의 경우 주문 단건 조회시 200 성공과 올바른 응답값이 오는지 확인합니다.")
@@ -238,7 +238,7 @@ class OrderSelectApiControllerTest {
 		given(orderService.getOrderWithNotDeleted(orderId)).willReturn(expectedOrderResponseDTO);
 		given(userDetailsService.loadUserByJwt(any())).willReturn(
 			new UserDetailsServiceImpl.CustomUserDetails(expectedOrderResponseDTO.getMember()));
-		
+
 		// when - then
 		ResultActions response = mockMvc.perform(get("/v1/api/orders/{orderId}", orderId)
 				.contentType(APPLICATION_JSON)
@@ -251,7 +251,7 @@ class OrderSelectApiControllerTest {
 			.andExpect(
 				jsonPath("$.member.email").value(expectedOrderResponseDTO.getMember().getEmail()))
 			.andExpect(jsonPath("$.member.role").value(Role.ROLE_CUSTOMER.name()));
-		
+
 		response
 			.andDo(MockMvcRestDocumentationWrapper.document(
 				"get-order-customer",
@@ -305,7 +305,7 @@ class OrderSelectApiControllerTest {
 				)
 			));
 	}
-	
+
 	@Test
 	@DisplayName("권한 - ADMIN 의 경우 주문 단건 조회시 200 성공과 올바른 응답값이 오는지 확인합니다.")
 	@MockMember(role = Role.ROLE_ADMIN)
@@ -316,7 +316,7 @@ class OrderSelectApiControllerTest {
 		given(orderService.getOrderWithNoCondition(orderId)).willReturn(expectedOrderResponseDTO);
 		given(userDetailsService.loadUserByJwt(any())).willReturn(
 			new UserDetailsServiceImpl.CustomUserDetails(expectedOrderResponseDTO.getMember()));
-		
+
 		// when - then
 		ResultActions response = mockMvc.perform(get("/v1/api/orders/{orderId}", orderId)
 				.contentType(APPLICATION_JSON)
@@ -330,7 +330,7 @@ class OrderSelectApiControllerTest {
 			.andExpect(
 				jsonPath("$.member.email").value(expectedOrderResponseDTO.getMember().getEmail()))
 			.andExpect(jsonPath("$.member.role").value(Role.ROLE_ADMIN.name()));
-		
+
 		response
 			.andDo(MockMvcRestDocumentationWrapper.document(
 				"get-order-customer",
@@ -384,7 +384,7 @@ class OrderSelectApiControllerTest {
 				)
 			));
 	}
-	
+
 	@Test
 	@MockMember(role = Role.ROLE_CUSTOMER)
 	@DisplayName("권한 - CUSTOMER 인 경우, 주문 단건 조회 실패 케이스의 경우 404 반환과 올바른 응답값이 오는지 확인합니다.")
@@ -395,14 +395,14 @@ class OrderSelectApiControllerTest {
 		given(userDetailsService.loadUserByJwt(any())).willReturn(
 			new UserDetailsServiceImpl.CustomUserDetails(
 				MemberResponseDTO.from(expectedOrderWithCUSTOMERUser.getMember())));
-		
+
 		// when - then
 		ResultActions response = mockMvc.perform(get("/v1/api/orders/{orderId}", orderId)
 				.contentType(APPLICATION_JSON)
 				.accept(APPLICATION_JSON))
 			.andExpect(status().isNotFound())
 			.andExpect(jsonPath("$.message").value(ORDER_NOT_FOUND.getMessage()));
-		
+
 		response
 			.andDo(MockMvcRestDocumentationWrapper.document(
 				"get-order-fail-customer",
@@ -424,7 +424,7 @@ class OrderSelectApiControllerTest {
 				)
 			));
 	}
-	
+
 	@Test
 	@MockMember(role = Role.ROLE_ADMIN)
 	@DisplayName("권한 - ADMIN 인 경우, 주문 단건 조회 실패 케이스의 경우 404 반환과 올바른 응답값이 오는지 확인합니다.")
@@ -434,14 +434,14 @@ class OrderSelectApiControllerTest {
 		given(orderService.getOrderWithNoCondition(orderId)).willThrow(new NotFoundException(ORDER_NOT_FOUND));
 		given(userDetailsService.loadUserByJwt(any())).willReturn(new UserDetailsServiceImpl.CustomUserDetails(
 			MemberResponseDTO.from(expectedOrderWithADMINUser.getMember())));
-		
+
 		// when - then
 		ResultActions response = mockMvc.perform(get("/v1/api/orders/{orderId}", orderId)
 				.contentType(APPLICATION_JSON)
 				.accept(APPLICATION_JSON))
 			.andExpect(status().isNotFound())
 			.andExpect(jsonPath("$.message").value(ORDER_NOT_FOUND.getMessage()));
-		
+
 		response
 			.andDo(MockMvcRestDocumentationWrapper.document(
 				"get-order-fail-admin",
@@ -463,7 +463,7 @@ class OrderSelectApiControllerTest {
 				)
 			));
 	}
-	
+
 	//다건 조회 테스트
 	@Test
 	@MockMember(role = Role.ROLE_CUSTOMER)
@@ -475,22 +475,22 @@ class OrderSelectApiControllerTest {
 		int pageNum = 0;
 		int pageSize = 10;
 		String requestYear = "2024";
-		
+
 		Pageable pageable = PageRequest.of(pageNum, pageSize);
-		
+
 		OrderResponseDTO expectedOrderResponseDTO = OrderResponseDTO.from(expectedOrderWithCUSTOMERUser);
 		expectedOrderResponseDTO.setDeletedDate(null);
-		
+
 		MemberResponseDTO memberDTO = MemberResponseDTO.builder().build();
-		
+
 		Page<OrderResponseDTO> expectedOrderResponseDTOPage = new PageImpl<>(List.of(expectedOrderResponseDTO), pageable, 1L);
-		
+
 		given(orderService.getOrdersWithPaginationAndNotDeleted(any(MemberEntity.class), any(OrderPageRequestDTO.class)))
 			.willReturn(expectedOrderResponseDTOPage);
-		
+
 		given(userDetailsService.loadUserByJwt(any()))
 			.willReturn(new UserDetailsServiceImpl.CustomUserDetails(memberDTO));
-		
+
 		// when - then
 		ResultActions response = mockMvc.perform(
 				get("/v1/api/orders/users/{userNumber}", userNumber)
@@ -501,7 +501,7 @@ class OrderSelectApiControllerTest {
 					.contentType(APPLICATION_JSON)
 					.accept(APPLICATION_JSON))
 			.andExpect(status().isOk());
-		
+
 		response
 			.andDo(MockMvcRestDocumentationWrapper.document(
 				"get-orders-customer",
@@ -577,7 +577,7 @@ class OrderSelectApiControllerTest {
 						fieldWithPath("size").description("페이지 크기")
 					).build())));
 	}
-	
+
 	// 404 NOT_FOUND_EXCEPTION 이 터지는 경우 테스트
 	@Test
 	@MockMember(role = Role.ROLE_CUSTOMER)
@@ -588,19 +588,19 @@ class OrderSelectApiControllerTest {
 		int pageNum = 100;
 		int pageSize = 0;
 		String requestYear = "2024";
-		
+
 		Member member = Member.from(MemberResponseDTO.from(expectedOrderWithCUSTOMERUser.getMember()));
-		
+
 		OrderPageRequestDTO orderPageRequestDTO = OrderPageRequestDTO.create(PageRequestDTO.create(pageNum, pageSize),
 			1L, requestYear);
-		
+
 		given(orderService.getOrdersWithPaginationAndNotDeleted(any(MemberEntity.class), any(OrderPageRequestDTO.class))).willThrow(
 			new NotFoundException(ORDER_NOT_FOUND));
-		
+
 		given(userDetailsService.loadUserByJwt(any())).willReturn(
 			new UserDetailsServiceImpl.CustomUserDetails(
 				MemberResponseDTO.from(expectedOrderWithCUSTOMERUser.getMember())));
-		
+
 		// when - then
 		ResultActions response = mockMvc.perform(
 				get("/v1/api/orders/users/{userNumber}", userNumber)
@@ -612,7 +612,7 @@ class OrderSelectApiControllerTest {
 					.accept(APPLICATION_JSON))
 			.andExpect(status().isNotFound())
 			.andExpect(jsonPath("$.message").value(ORDER_NOT_FOUND.getMessage()));
-		
+
 		response
 			.andDo(MockMvcRestDocumentationWrapper.document(
 				"get-orders-fail-customer",
@@ -634,7 +634,7 @@ class OrderSelectApiControllerTest {
 				)
 			));
 	}
-	
+
 	@Test
 	@MockMember(role = Role.ROLE_ADMIN)
 	@DisplayName("권한 - ADMIN 인 경우, 삭제되지 않은 주문을 조회하는 서비스를 호출 시 200 성공을 기대합니다.")
@@ -644,22 +644,22 @@ class OrderSelectApiControllerTest {
 		int pageNum = 0;
 		int pageSize = 10;
 		String requestYear = "2024";
-		
+
 		Pageable pageable = PageRequest.of(pageNum, pageSize);
-		
+
 		OrderResponseDTO expectedOrderResponseDTO = OrderResponseDTO.from(expectedOrderWithADMINUser);
 		expectedOrderResponseDTO.setDeletedDate(null);
-		
+
 		MemberResponseDTO memberDTO = MemberResponseDTO.builder().build();
-		
+
 		Page<OrderResponseDTO> expectedOrderResponseDTOPage = new PageImpl<>(List.of(expectedOrderResponseDTO), pageable, 1L);
-		
+
 		given(orderService.getOrdersWithPaginationAndNotDeleted(any(MemberEntity.class), any(OrderPageRequestDTO.class)))
 			.willReturn(expectedOrderResponseDTOPage);
-		
+
 		given(userDetailsService.loadUserByJwt(any()))
 			.willReturn(new UserDetailsServiceImpl.CustomUserDetails(memberDTO));
-		
+
 		// when - then
 		ResultActions response = mockMvc.perform(
 				get("/v1/api/orders/users/{userNumber}", userNumber)
@@ -670,7 +670,7 @@ class OrderSelectApiControllerTest {
 					.contentType(APPLICATION_JSON)
 					.accept(APPLICATION_JSON))
 			.andExpect(status().isOk());
-		
+
 		response
 			.andDo(MockMvcRestDocumentationWrapper.document(
 				"get-orders-admin",
@@ -746,7 +746,7 @@ class OrderSelectApiControllerTest {
 						fieldWithPath("size").description("페이지 크기")
 					).build())));
 	}
-	
+
 	@Test
 	@MockMember(role = Role.ROLE_ADMIN)
 	@DisplayName("권한 - ADMIN 인 경우, 주문 다건 조회 실패 케이스의 경우 404 반환과 올바른 응답값이 오는지 확인합니다.")
@@ -756,19 +756,19 @@ class OrderSelectApiControllerTest {
 		int pageNum = 100;
 		int pageSize = 0;
 		String requestYear = "2024";
-		
+
 		OrderPageRequestDTO orderPageRequestDTO = OrderPageRequestDTO.create(PageRequestDTO.create(pageNum, pageSize),
 			1L, requestYear);
-		
+
 		Member member = Member.from(MemberResponseDTO.from(expectedOrderWithADMINUser.getMember()));
-		
+
 		given(orderService.getOrdersWithPaginationAndNoCondition(any(MemberEntity.class), any(OrderPageRequestDTO.class))).willThrow(
 			new NotFoundException(ORDER_NOT_FOUND));
-		
+
 		given(userDetailsService.loadUserByJwt(any())).willReturn(
 			new UserDetailsServiceImpl.CustomUserDetails(
 				MemberResponseDTO.from(expectedOrderWithADMINUser.getMember())));
-		
+
 		// when - then
 		ResultActions response = mockMvc.perform(
 				get("/v1/api/orders/users/{userNumber}", userNumber)
@@ -780,7 +780,7 @@ class OrderSelectApiControllerTest {
 					.accept(APPLICATION_JSON))
 			.andExpect(status().isNotFound())
 			.andExpect(jsonPath("$.message").value(ORDER_NOT_FOUND.getMessage()));
-		
+
 		response
 			.andDo(MockMvcRestDocumentationWrapper.document(
 				"get-orders-fail-admin",
