+++ src/test/java/com/psj/itembrowser/order/service/OrderSelectWithDBServiceTest.java
@@ -34,13 +34,9 @@ import com.psj.itembrowser.order.mapper.OrderMapper;
 import com.psj.itembrowser.order.persistence.OrderPersistence;
 import com.psj.itembrowser.order.repository.CustomOrderRepository;
 import com.psj.itembrowser.order.repository.OrderRepository;
-import com.psj.itembrowser.order.service.impl.OrderCalculationService;
-import com.psj.itembrowser.order.service.impl.OrderService;
-import com.psj.itembrowser.order.service.impl.PaymentService;
-import com.psj.itembrowser.order.service.impl.ShppingInfoValidationService;
 import com.psj.itembrowser.product.domain.entity.ProductEntity;
-import com.psj.itembrowser.product.service.impl.ProductService;
-import com.psj.itembrowser.product.service.impl.ProductValidationHelper;
+import com.psj.itembrowser.product.service.ProductService;
+import com.psj.itembrowser.product.service.ProductValidationHelper;
 import com.psj.itembrowser.security.auth.service.impl.AuthenticationService;
 import com.psj.itembrowser.security.common.exception.NotFoundException;
 import com.psj.itembrowser.shippingInfos.domain.entity.ShippingInfoEntity;
@@ -48,76 +44,76 @@ import com.psj.itembrowser.shippingInfos.domain.entity.ShippingInfoEntity;
 @ActiveProfiles("test")
 @RepositoryTest
 public class OrderSelectWithDBServiceTest {
-	
+
 	@Autowired
 	private OrderRepository orderRepository;
-	
+
 	@Autowired
 	private MemberRepository memberRepository;
-	
+
 	@Autowired
 	private CustomOrderRepository customOrderRepository;
-	
+
 	@MockBean
 	private OrderMapper orderMapper;
-	
+
 	@Mock
 	private AuthenticationService authenticationService;
-	
+
 	@Mock
 	private PaymentService paymentService;
-	
+
 	@Mock
 	private OrderCalculationService orderCalculationService;
-	
+
 	@Mock
 	private ProductValidationHelper productValidationHelper;
-	
+
 	@Mock
 	private ShppingInfoValidationService shppingInfoValidationService;
-	
+
 	@Mock
 	private ProductService productService;
-	
+
 	private OrderService orderService;
-	
+
 	@PersistenceContext
 	EntityManager em;
-	
+
 	private Long validOrderId;
-	
+
 	private Long invalidOrderId;
-	
+
 	private OrderEntity validOrder;
-	
+
 	private MemberEntity savedMember;
-	
+
 	@BeforeEach
 	void init() {
 		OrderPersistence orderPersistence = new OrderPersistence(orderMapper, orderRepository, customOrderRepository);
 		orderService = new OrderService(orderPersistence, orderMapper, orderCalculationService, authenticationService, productValidationHelper,
 			shppingInfoValidationService, paymentService, productService);
 	}
-	
+
 	@BeforeEach
 	public void setUp() {
 		validOrderId = 1L;
 		invalidOrderId = 99L;
-		
+
 		MemberEntity expectedMember = MemberEntity.builder()
 			.credentials(Credentials.builder().email("qkrtkdwns3410@gmail.com").password("PasswordIsHard!@#").build())
 			.address(Address.builder().addressMain("포항시 남구 연일읍 유강길 10 - 44").addressSub("김밥아파트 101동 302호").zipCode("08593").build())
 			.name(Name.builder().firstName("홍").lastName("길동").build())
 			.build();
-		
+
 		savedMember = memberRepository.save(expectedMember);
-		
+
 		ShippingInfoEntity expectedShippingInfoEntity = ShippingInfoEntity.builder()
 			.memberNo(savedMember.getMemberNo())
 			.build();
-		
+
 		em.persist(expectedShippingInfoEntity);
-		
+
 		OrdersProductRelationEntity expectedOrderRelation = OrdersProductRelationEntity.builder()
 			.groupId(1L)
 			.productId(1L)
@@ -126,7 +122,7 @@ public class OrderSelectWithDBServiceTest {
 			.updatedDate(LocalDateTime.now())
 			.deletedDate(null)
 			.build();
-		
+
 		this.validOrder = OrderEntity.builder()
 			.id(1L)
 			.member(expectedMember)
@@ -138,24 +134,24 @@ public class OrderSelectWithDBServiceTest {
 			.updatedDate(LocalDateTime.now())
 			.deletedDate(null)
 			.build();
-		
+
 		ProductEntity productEntity = ProductEntity.builder().name("섬유유연제").unitPrice(1000).quantity(10).build();
-		
+
 		em.persist(productEntity);
-		
+
 		em.flush();
 	}
-	
+
 	@Test
 	@DisplayName("조건 없이 주문 조회 후 주문 정보 반환이 올바르게 되는지 테스트")
 	void When_GetOrderWithNoCondition_Expect_ReturnOrderResponseDTO() {
 		//given
-		
+
 		OrderEntity saved = orderRepository.save(validOrder);
-		
+
 		//when
 		OrderResponseDTO orderResponseDTO = orderService.getOrderWithNoCondition(saved.getId());
-		
+
 		//then
 		assertThat(orderResponseDTO).isNotNull();
 		assertThat(orderResponseDTO.getId()).isEqualTo(saved.getId());
@@ -163,16 +159,16 @@ public class OrderSelectWithDBServiceTest {
 		assertThat(orderResponseDTO.getOrdersProductRelations()).hasSize(1);
 		assertThat(orderResponseDTO.getOrdersProductRelations().get(0).getProductId()).isEqualTo(1L);
 	}
-	
+
 	@Test
 	@DisplayName("삭제되지 않은 주문 조회 후 주문 정보 반환이 올바르게 되는지 테스트")
 	void When_GetOrderWithNotDeleted_Expect_ReturnOrderResponseDTO() {
 		//given
 		OrderEntity saved = orderRepository.save(validOrder);
-		
+
 		//when
 		OrderResponseDTO orderResponseDTO = orderService.getOrderWithNotDeleted(saved.getId());
-		
+
 		//then
 		assertThat(orderResponseDTO).isNotNull();
 		assertThat(orderResponseDTO.getId()).isEqualTo(saved.getId());
@@ -180,16 +176,16 @@ public class OrderSelectWithDBServiceTest {
 		assertThat(orderResponseDTO.getOrdersProductRelations()).hasSize(1);
 		assertThat(orderResponseDTO.getOrdersProductRelations().get(0).getProductId()).isEqualTo(1L);
 	}
-	
+
 	@Test
 	@DisplayName("주문에 대한 단건 조회의 경우 메서드가 정상적으로 orderResponseDTO 를 반환하는지")
 	void test() {
 		//given
 		OrderEntity saved = orderRepository.save(validOrder);
-		
+
 		//when
 		OrderResponseDTO orderResponseDTO = orderService.getOrderWithNoCondition(saved.getId());
-		
+
 		//then
 		assertThat(orderResponseDTO).isNotNull();
 		assertThat(orderResponseDTO.getId()).isEqualTo(saved.getId());
@@ -197,7 +193,7 @@ public class OrderSelectWithDBServiceTest {
 		assertThat(orderResponseDTO.getOrdersProductRelations()).hasSize(1);
 		assertThat(orderResponseDTO.getOrdersProductRelations().get(0).getProductId()).isEqualTo(1L);
 	}
-	
+
 	@Test
 	@DisplayName("조건 없이 주문 조회 시 주문 정보가 없을 경우 NotFoundException 발생")
 	void When_GetOrderWithNoCondition_Expect_ThrowNotFoundException() {
@@ -206,7 +202,7 @@ public class OrderSelectWithDBServiceTest {
 			.isInstanceOf(NotFoundException.class)
 			.hasMessageContaining("Not Found Order");
 	}
-	
+
 	@Test
 	@DisplayName("삭제되지 않은 주문 조회 시 주문 정보가 없을 경우 NotFoundException 발생")
 	void When_GetOrder_Expect_ThrowNotFoundException() {
@@ -214,20 +210,20 @@ public class OrderSelectWithDBServiceTest {
 		assertThatThrownBy(() -> orderService.getOrderWithNotDeleted(invalidOrderId))
 			.isInstanceOf(NotFoundException.class).hasMessageContaining("Not Found Order");
 	}
-	
+
 	@Test
 	@DisplayName("다건 주문 조회 (getOrdersWithPaginationAndNoCondition) - 모든 정보 조회시 주문 정보가 있을 경우 주문 정보 리스트 반환")
 	void When_GetOrdersWithPaginationAndNoCondition_Expect_ReturnOrderResponseDTOList() {
 		//given
 		MemberEntity member = MemberEntity.builder().role(Role.ROLE_CUSTOMER).build();
-		
+
 		OrderEntity expectedOrder = orderRepository.save(validOrder);
-		
+
 		OrderPageRequestDTO dto = OrderPageRequestDTO.builder().pageNum(0).pageSize(1).userNumber(1L).build();
-		
+
 		//when
 		Page<OrderResponseDTO> ordersWithPaginationAndNoCondition = orderService.getOrdersWithPaginationAndNoCondition(member, dto);
-		
+
 		//then
 		assertThat(ordersWithPaginationAndNoCondition).isNotNull();
 		assertThat(ordersWithPaginationAndNoCondition.getContent()).isNotNull();
@@ -237,21 +233,21 @@ public class OrderSelectWithDBServiceTest {
 		assertThat(ordersWithPaginationAndNoCondition.getContent().get(0).getOrdersProductRelations()).hasSize(1);
 		assertThat(ordersWithPaginationAndNoCondition.getContent().get(0).getOrdersProductRelations().get(0).getProductId()).isEqualTo(1L);
 	}
-	
+
 	@Test
 	@DisplayName("다건 주문 조회 (getOrdersWithPaginationAndNoCondition) - 모든 정보 조회시 주문 정보가 없을 경우 NotFoundException 발생")
 	void When_GetOrdersWithPaginationAndNoCondition_Expect_ThrowNotFoundException() {
 		//given
 		MemberEntity member = MemberEntity.builder().role(Role.ROLE_CUSTOMER).build();
-		
+
 		OrderPageRequestDTO dto = OrderPageRequestDTO.builder().pageNum(0).pageSize(1).userNumber(1L).build();
-		
+
 		//when - then
 		assertThatThrownBy(() -> orderService.getOrdersWithPaginationAndNoCondition(member, dto))
 			.isInstanceOf(NotFoundException.class)
 			.hasMessageContaining("Not Found Order");
 	}
-	
+
 	@Test
 	@DisplayName("다건 주문 조회 (getOrdersWithPaginationAndNotDeleted) - 삭제되지 않은 주문 조회시 주문 정보가 있을 경우 주문 정보 리스트 반환")
 	void When_GetOrdersWithPaginationAndNotDeleted_Expect_ReturnOrderResponseDTOList() {
@@ -263,9 +259,9 @@ public class OrderSelectWithDBServiceTest {
 			.name(Name.builder().firstName("홍").lastName("길동").build())
 			.role(Role.ROLE_CUSTOMER)
 			.build();
-		
+
 		MemberEntity savedMember = memberRepository.save(member);
-		
+
 		OrderEntity deletedOrder = OrderEntity.builder()
 			.id(1L)
 			.orderStatus(OrderStatus.CANCELED)
@@ -274,9 +270,9 @@ public class OrderSelectWithDBServiceTest {
 			.createdDate(LocalDateTime.now().minusDays(1))
 			.updatedDate(LocalDateTime.now().minusDays(1))
 			.build();
-		
+
 		orderRepository.save(deletedOrder);
-		
+
 		OrderEntity notDeletedOrder = OrderEntity.builder()
 			.id(2L)
 			.orderStatus(OrderStatus.ACCEPT)
@@ -285,16 +281,16 @@ public class OrderSelectWithDBServiceTest {
 			.createdDate(LocalDateTime.now().minusDays(1))
 			.updatedDate(LocalDateTime.now().minusDays(1))
 			.build();
-		
+
 		OrderEntity expectedOrder = orderRepository.save(notDeletedOrder);
-		
+
 		OrderPageRequestDTO dto = OrderPageRequestDTO.builder().pageNum(0).pageSize(10).userNumber(savedMember.getMemberNo()).build();
-		
+
 		doNothing().when(authenticationService).authorizeOrdersWhenCustomer(any(Page.class), any(MemberEntity.class));
-		
+
 		//when
 		Page<OrderResponseDTO> ordersWithPaginationAndNotDeleted = orderService.getOrdersWithPaginationAndNotDeleted(member, dto);
-		
+
 		//then
 		assertThat(ordersWithPaginationAndNotDeleted).isNotNull();
 		assertThat(ordersWithPaginationAndNotDeleted.getContent()).isNotNull();
@@ -305,15 +301,15 @@ public class OrderSelectWithDBServiceTest {
 		assertThat(ordersWithPaginationAndNotDeleted.getContent().get(0).getOrderStatus()).isEqualTo(expectedOrder.getOrderStatus());
 		assertThat(ordersWithPaginationAndNotDeleted.getContent().get(0).getOrdersProductRelations()).hasSize(0);
 	}
-	
+
 	@Test
 	@DisplayName("삭제되지 않은 주문 조회 시 주문 정보가 없을 경우 NotFoundException 발생")
 	void When_GetOrdersWithPaginationAndNotDeleted_Expect_ThrowNotFoundException() {
 		//given
 		MemberEntity member = MemberEntity.builder().role(Role.ROLE_CUSTOMER).build();
-		
+
 		OrderPageRequestDTO dto = OrderPageRequestDTO.builder().pageNum(0).pageSize(1).userNumber(1L).build();
-		
+
 		//when - then
 		assertThatThrownBy(() -> orderService.getOrdersWithPaginationAndNotDeleted(member, dto))
 			.isInstanceOf(NotFoundException.class)
